requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.4.0
constantly==23.10.4
cryptography==43.0.3
daphne==4.1.2
defusedxml==0.8.0rc2
Django==5.1.3
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
oauthlib==3.2.2
phonenumbers==8.13.49
pillow==11.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
redis==5.2.0
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==75.4.0
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
zope.interface==7.1.1
